+++ src/main/java/com/jcloisterzone/ui/grid/layer/AbstractAreaLayer.java
@@ -166,9 +166,11 @@ public abstract class AbstractAreaLayer extends AbstractGridLayer implements Gri
     /** standard highlight **/
     private void paintAreaHighlight(Graphics2D g2) {
         Player p = getClient().getGame().getActivePlayer();
-        g2.setColor(p.getColors().getMeepleColor());
-        g2.setComposite(AREA_ALPHA_COMPOSITE);
-        g2.fill(transformArea(areas.get(selectedLocation), selectedPosition));
+        if (p != null) { //sync issue
+            g2.setColor(p.getColors().getMeepleColor());
+            g2.setComposite(AREA_ALPHA_COMPOSITE);
+            g2.fill(transformArea(areas.get(selectedLocation), selectedPosition));
+        }
     }
 
     @Override